Former-commit-id: 3cedfb38d77c060b6b886891c2213b49c1faf52f

+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Commands/Interview/Base/QuestionCommand.cs
@@ -5,9 +5,9 @@ namespace WB.Core.SharedKernels.DataCollection.Commands.Interview.Base
     public abstract class QuestionCommand : InterviewCommand
     {
         public Guid QuestionId { get; private set; }
-        public int[] RosterVector { get; private set; }
+        public decimal[] RosterVector { get; private set; }
 
-        protected QuestionCommand(Guid interviewId, Guid userId, Guid questionId, int[] rosterVector)
+        protected QuestionCommand(Guid interviewId, Guid userId, Guid questionId, decimal[] rosterVector)
             : base(interviewId, userId)
         {
             this.QuestionId = questionId;
